5th commit
-Изменено расположение классов контроллера и сервиса
-Добавлена аннотация @RestController к контроллерам
-В функциях findStudent и findFaculty @GetMapping заменена на @PutMapping
-Функции updateStudent и updateFaculty Теперь принримают только один параметр
-В функции deleteStudent и deleteFaculty @GetMapping заменена на @DeleteMapping
-В функциях colorSearch и ageSearch @RequestBody заменена на @RequestParam
-В классах Faculty и Student изменен метод equals()
-В функциях createFaculty и createStudent удалена проверка на наличие объекта
